+++ CHANGELOG.md
@@ -1222,3 +1222,5 @@ All notable changes to this project will be documented in this file.
 - fix beautify indent on shorthand if else
 - make installer code more verbose, including more error messages
 - update message-hook agent to version 0.5, being able to properly sync windows between the two clients (NOTE: you'll need download the newest dll manually)
+- add postCommand option to build
+- only expose one executable greybel command and instead use nested commands for different functionalities
